move test hooks
