+++ spec/hooks/useQuizState/useQuizApiState/useQuizApiState.test.tsx
@@ -1,8 +1,8 @@
 import { renderHook, waitFor } from '@testing-library/react';
-import { mockConstructorIOClient } from '../../__tests__/utils';
-import useQuizApiState from '../../../src/hooks/useQuizState/useQuizApiState';
-import { getQuizResults } from '../../../src/services';
-import { QUIZ_VERSION_ID, QUIZ_ID } from '../../__tests__/constants';
+import { mockConstructorIOClient } from '../../../__tests__/utils';
+import useQuizApiState from '../../../../src/hooks/useQuizState/useQuizApiState';
+import { getQuizResults } from '../../../../src/services';
+import { QUIZ_VERSION_ID, QUIZ_ID } from '../../../__tests__/constants';
 
 jest.mock('../../../src/services', () => ({
   getNextQuestion: jest.fn().mockResolvedValue({
